// src/enums/item.ts
export enum EItemType {
    Natural = 'Material',
    Products = 'Products',
    Usables = 'Usables',
}

export enum EItemCategory {
    Ore = 'Ore',
    Plant = 'Plant',
    Refined = 'Refined',
    Powder = 'Powder',
    Bottle = 'Bottle',
    Parts = 'Parts',
    Component = 'Component',
    Battery = 'Battery',
    Solution = 'Solution',
    Consumable = 'Consumable',
    Other = 'Other',
}